+++ VTK/Patented/vtkKitwareCutter.cxx
@@ -124,7 +124,7 @@ void vtkKitwareCutter::StructuredPointsCutter()
   
   
   int i;
-  float scalar;
+  double scalar;
   for (i = 0; i < numPts; i++)
     {
     scalar = this->CutFunction->FunctionValue(input->GetPoint(i));
@@ -184,7 +184,7 @@ void vtkKitwareCutter::StructuredGridCutter()
     }
   
   int i;
-  float scalar;
+  double scalar;
   for (i = 0; i < numPts; i++)
     {
     scalar = this->CutFunction->FunctionValue(input->GetPoint(i));
@@ -243,7 +243,7 @@ void vtkKitwareCutter::RectilinearGridCutter()
     }
   
   int i;
-  float scalar;
+  double scalar;
   for (i = 0; i < numPts; i++)
     {
     scalar = this->CutFunction->FunctionValue(input->GetPoint(i));
